% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqNjoy.R
\name{SeqNjoy}
\alias{SeqNjoy}
\title{Start SeqNjoy app}
\usage{
SeqNjoy()
}
\value{
either run the app as a side effect or return a shiny.appobj object
}
\description{
This function starts the SeqNjoy server and launches the browser
}
\examples{
# This example runs only in interactive sessions:
if (interactive()) {
    SeqNjoy()
}

}
